# api/swagger/v1/catalog.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Catalog service",
    "version": "1.0",
    "contact": {
      "name": "catalog service",
      "url": "https://github.com/ckbball/smurfin-catalog",
      "email": "blah@happy.org"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/catalogs": {
      "post": {
        "operationId": "Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/catalogv1CreateResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/catalogv1CreateRequest"
            }
          }
        ],
        "tags": [
          "CatalogService"
        ]
      }
    },
    "/v1/catalogs/list": {
      "post": {
        "operationId": "ListItems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/catalogv1ListResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/catalogv1ListRequest"
            }
          }
        ],
        "tags": [
          "CatalogService"
        ]
      }
    },
    "/v1/catalogs/search": {
      "post": {
        "operationId": "FindItems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/catalogv1Response"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/catalogv1Specification"
            }
          }
        ],
        "tags": [
          "CatalogService"
        ]
      }
    },
    "/v1/catalogs/{id}": {
      "delete": {
        "operationId": "RemoveItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/catalogv1RemoveResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CatalogService"
        ]
      }
    }
  },
  "definitions": {
    "catalogv1CreateRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "item": {
          "$ref": "#/definitions/catalogv1Item"
        }
      }
    },
    "catalogv1CreateResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "catalogv1Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "vendor_id": {
          "type": "string"
        },
        "champions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "blue_essence": {
          "type": "integer",
          "format": "int32"
        },
        "riot_points": {
          "type": "integer",
          "format": "int32"
        },
        "solo": {
          "type": "integer",
          "format": "int32"
        },
        "flex": {
          "type": "integer",
          "format": "int32"
        },
        "price_dollars": {
          "type": "integer",
          "format": "int32"
        },
        "price_cents": {
          "type": "integer",
          "format": "int32"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "login_password": {
          "type": "string"
        }
      }
    },
    "catalogv1ListRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "page": {
          "type": "string",
          "format": "int64"
        },
        "limit": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "catalogv1ListResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/catalogv1Item"
          }
        }
      }
    },
    "catalogv1RemoveResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "deleted": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "catalogv1Response": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/catalogv1Item"
          }
        }
      }
    },
    "catalogv1Specification": {
      "type": "object",
      "properties": {
        "solo": {
          "type": "integer",
          "format": "int32"
        },
        "flex": {
          "type": "integer",
          "format": "int32"
        },
        "price_dollars": {
          "type": "integer",
          "format": "int32"
        },
        "price_cents": {
          "type": "integer",
          "format": "int32"
        },
        "item_id": {
          "type": "string"
        },
        "vendor_id": {
          "type": "string"
        },
        "page_num": {
          "type": "integer",
          "format": "int32"
        },
        "item": {
          "$ref": "#/definitions/catalogv1Item"
        },
        "api": {
          "type": "string"
        }
      }
    }
  }
}